java/src/main/java/dojo/supermarket/model/Product.java
package dojo.supermarket.model;

import java.util.Objects;

public class Product {
    private final String name;
    private final ProductUnitType unit;

    public Product(String name, ProductUnitType unit) {
        this.name = name;
        this.unit = unit;
    }

    public String getName() {
        return name;
    }


    public ProductUnitType getUnit() {
        return unit;
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;
        Product product = (Product) o;
        return Objects.equals(name, product.name) &&
                unit == product.unit;
    }

    @Override
    public int hashCode() {

        return Objects.hash(name, unit);
    }
}
